pyproject.toml
[tool.poetry]
name = "betano-api"
version = "0.1.0"
description = ""
authors = ["Marcos Pereira"]
license = "MIT"
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "3.12.5"
notebook = "^7.2.2"
numpy = "^2.1.2"
pandas = "^2.2.3"
undetected-chromedriver = "^3.5.5"
ipython = "^8.28.0"
fastapi = {extras = ["standard"], version = "^0.115.2"}
email-validator = "^2.2.0"
chromedriver-py = "^129.0.6668.100"
pytest-playwright = "^0.5.2"

[tool.poetry.group.dev.dependencies]
mypy = "^1.12.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
